Renamed main file
change from web.py to temperature-monitor.py. No functionality change
